+++ src/bluesearch/mining/__init__.py
@@ -18,12 +18,7 @@
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 from .attribute import AttributeAnnotationTab, AttributeExtractor, TextCollectionWidget
-from .entity import (
-    PatternCreator,
-    check_patterns_agree,
-    global2model_patterns,
-    remap_entity_type,
-)
+from .entity import PatternCreator, check_patterns_agree, global2model_patterns
 from .eval import annotations2df, spacy2df
 from .pipeline import SPECS, run_pipeline
 from .relation import ChemProt, REModel, StartWithTheSameLetter, annotate
@@ -41,7 +36,6 @@ __all__ = [
     "annotations2df",
     "check_patterns_agree",
     "global2model_patterns",
-    "remap_entity_type",
     "run_pipeline",
     "spacy2df",
 ]